+++ framework/Runners/Factory.py
@@ -25,6 +25,7 @@ if not 'xrange' in dir(__builtins__):
 #End compatibility block for Python 3-------------------------------------------
 
 ################################################################################
+from utils import utils
 from .Runner import Runner
 from .ExternalRunner import ExternalRunner
 from .InternalRunner import InternalRunner
@@ -43,7 +44,7 @@ from .InternalThreadedRunner import InternalThreadedRunner
 __base = 'Runner'
 __interFaceDict = {}
 
-for classObj in eval(__base).__subclasses__():
+for classObj in utils.getAllSubclasses(eval(__base)):
   key = classObj.__name__
   __interFaceDict[key] = classObj
