+++ framework/utils/utils.py
@@ -1106,3 +1106,11 @@ def returnIdSeparator():
     @ Out, __idSeparator, string, the id separator
   """
   return __idSeparator
+
+def getAllSubclasses(cls):
+  """
+    Recursively collect all of the classes that are a subclass of cls
+    @ In, cls, the class to retrieve sub-classes.
+    @ Out, getAllSubclasses, list of class objects for each subclass of cls.
+  """
+  return cls.__subclasses__() + [g for s in cls.__subclasses__() for g in getAllSubclasses(s)]
